Create gtranslate_block_machine_translations.php
This was an attempt to block the machine translations that GTranslate automatically makes for all pages. The function was called in the template's index.php file, after the menu.

Unfortunately, due to cache problems the current language code wasn't always correctly grabbed. We weren't able to solve this with cache exclusion of GTranslate.